Change world name detection, drop alpha 17 compatibility
Since dedicated and sp world generation are different, I changed
how the world name is detected so I can generate the seed in sp,
and save the preview. That makes it incompatible with alpha 17,
I believe.

There's also another subtle change that makes it possible to generate
multiple worlds in sp, and save each one. Save preview will always
save the _last_ world that was generated.

One possible improvement would be to save _all_ worlds in the log,
or even to simply save all worlds in the GeneratedWorlds folder, and
avoid having to look it up on the log altogether. I think the main
reason that's not how it works presently is that it helped me detect
user (me) errors, both operation and coding.