Print ordinal classification of seeds, and remove -SIZE
The ordinal classification (1st, 2nd, 3rd, etc, but without the
suffix) is the first step towards highlighting all "1st", regardless
or how many are tied for that position.

It also helps evaluate how far down the list on other criterias a
seed is, and makes it possible to just grep the seed on a list and
know what it's position is.

To make up for the extra space, I'm removing -SIZE (-8192, -4096)
when printing. That also makes it easier to copy seed names.
Printing sizes predates things like "prefabs.sh" that already
extracts seeds into separate folders by size, and was done at a
time when I was actively interested in comparing seeds of
different sizes.

With the new format, each column is actually one character shorter
then they used to be, which is rather welcome, tbh. I've been
toying with the idea of switching seed generation to something that
would favor shorter names by using the 1/2 syllable dictionaries,
and/or rejecting seeds above a certain length.